fix: sales jpa services
